Extract meme sorts into module
