+++ routes/index.js
@@ -4,24 +4,8 @@ const router = express.Router();
 const { User, Meme, Comment, Like, Follow } = require("../db/models");
 const { logoutUser } = require("../auth");
 const { asyncHandler, isLoggedIn } = require("../utils");
+const { memesByComments, memesByLikes } = require("./utils/meme-sorts");
 
-const memesByComments = (a, b) => {
-  if (a.comments.length < b.comments.length) {
-    return 1;
-  } else if (a.comments.length > b.comments.length) {
-    return -1;
-  }
-  return 0;
-};
-
-const memesByLikes = (a, b) => {
-  if (a.likes.length < b.likes.length) {
-    return 1;
-  } else if (a.likes.length > b.likes.length) {
-    return -1;
-  }
-  return 0;
-};
 /* GET home page -- default sorted by likes. */
 router.get(
   "/",
@@ -48,22 +32,24 @@ router.get(
       .sort((a, b) => memesByLikes(a, b))
       .slice(0, 10);
 
-    const isLoggedIn = (req) => req.session.auth === undefined;
     // if user logged in, render landing-page, else render index
     res.render(isLoggedIn(req) ? "landing-page" : "index", {
       title: "Memehub",
       trendingMemes,
       bestMemes,
       i: 1,
-      j: 1,
     });
   })
 );
 
 // GET home page sorted by recent
-router.get("/recent", (req, res, next) => {
-  res.render("index", { title: "Memehub" });
-});
+router.get(
+  "/recent",
+  asyncHandler(async (req, res, next) => {
+    const latestMemes = await Meme.findAll({ order: [["id", "DESC"]] });
+    res.render("index", { title: "Memehub", latestMemes });
+  })
+);
 
 // GET home page sorted by most comments
 router.get("/hot", (req, res, next) => {
